[REFACTOR] 중복 예외 처리
