+++ src/main/java/com/dailyon/productservice/exception/UniqueException.java
@@ -1,6 +1,9 @@
 package com.dailyon.productservice.exception;
 
 public class UniqueException extends RuntimeException {
+    public static String DUPLICATE_PRODUCT_SIZE_NAME = "이미 존재하는 치수값입니다";
+    public static String DUPLICATE_BRAND_NAME = "이미 존재하는 브랜드 이름입니다";
+    public static String DUPLICATE_CATEGORY_NAME = "이미 존재하는 카테고리 이름입니다";
     public UniqueException() {
         super();
     }
