+++ src/main/java/com/dailyon/productservice/exception/advice/GlobalControllerAdvice.java
@@ -1,6 +1,8 @@
 package com.dailyon.productservice.exception.advice;
 
+import com.dailyon.productservice.exception.NotExistsException;
 import com.dailyon.productservice.exception.ErrorResponse;
+import com.dailyon.productservice.exception.UniqueException;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -14,6 +16,22 @@ public class GlobalControllerAdvice {
     public ErrorResponse validException(BindException e) {
         return ErrorResponse.builder()
                 .message(e.getBindingResult().getFieldError().getDefaultMessage())
+                .build();
+    }
+
+    @ExceptionHandler(NotExistsException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse notExistsException(NotExistsException e) {
+        return ErrorResponse.builder()
+                .message(e.getMessage())
+                .build();
+    }
+
+    @ExceptionHandler(UniqueException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse duplicatedCategoryException(UniqueException e) {
+        return ErrorResponse.builder()
+                .message(e.getMessage())
                 .build();
     }
 }
