--- src/main/java/com/dailyon/productservice/exception/advice/admin/BrandAdminControllerAdvice.java
@@ -1,29 +1,0 @@
-package com.dailyon.productservice.exception.advice.admin;
-
-import com.dailyon.productservice.controller.admin.BrandAdminController;
-import com.dailyon.productservice.exception.NotExistsException;
-import com.dailyon.productservice.exception.UniqueException;
-import com.dailyon.productservice.exception.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice(basePackageClasses = BrandAdminController.class)
-public class BrandAdminControllerAdvice {
-    @ExceptionHandler(UniqueException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse duplicatedBrandException(UniqueException e) {
-        return ErrorResponse.builder()
-                .message("이미 존재하는 브랜드 이름입니다")
-                .build();
-    }
-
-    @ExceptionHandler(NotExistsException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse brandNotFoundException(NotExistsException e) {
-        return ErrorResponse.builder()
-                .message(e.getMessage())
-                .build();
-    }
-}
